refactor(auth): split snapshot and replica
